+++ src/electionguard-ui/ElectionGuard.UI.Lib/Models/ElectionPartialKeyVerification.cs
@@ -3,11 +3,16 @@
 /// <summary>
 /// Verification of election partial key used in key sharing
 /// </summary>
-public record ElectionPartialKeyVerification
+public class ElectionPartialKeyVerification : DatabaseRecord
 {
     public string? KeyCeremonyId { get; set; }
     public string? OwnerId { get; init; }
     public string? DesignatedId { get; init; }
     public string? VerifierId { get; init; }
     public bool Verified { get; init; }
+
+    public ElectionPartialKeyVerification() : base(nameof(ElectionPartialKeyVerification))
+    {
+
+    }
 }
