+++ src/electionguard-ui/ElectionGuard.UI.Lib/Models/GuardianPublicKey.cs
@@ -1,0 +1,14 @@
+﻿namespace ElectionGuard.UI.Lib.Models;
+
+public class GuardianPublicKey : DatabaseRecord
+{
+    public string? KeyCeremonyId { get; set; }
+
+    public string? GuardianId { get; set; }
+
+    public ElectionPublicKey? PublicKey { get; set; }
+
+    public GuardianPublicKey() : base(nameof(GuardianPublicKey))
+    {
+    }
+}
