+++ src/electionguard-ui/ElectionGuard.UI.Lib/Models/KeyCeremony.cs
@@ -3,61 +3,8 @@
 namespace ElectionGuard.UI.Lib.Models;
 
 /// <summary>
-/// The Election joint key
+/// Data for the key ceremony that is saved to the database
 /// </summary>
-public class ElectionJointKey : DisposableBase
-{
-    /// <summary>
-    /// The product of the guardian public keys
-    /// K = ∏ ni=1 Ki mod p.
-    /// </summary>
-    public ElementModP? JointPublicKey { get; init; }
-
-    /// <summary>
-    /// The hash of the commitments that the guardians make to each other
-    /// H = H(K 1,0 , K 2,0 ... , K n,0 )
-    /// </summary>
-    public ElementModQ? CommitmentHash { get; init; }
-
-    protected override void DisposeUnmanaged()
-    {
-        base.DisposeUnmanaged();
-
-        JointPublicKey?.Dispose();
-        CommitmentHash?.Dispose();
-    }
-}
-
-public class GuardianPublicKey : DatabaseRecord
-{
-    public string? KeyCeremonyId { get; set; }
-
-    public string? GuardianId { get; set; }
-
-    public ElectionPublicKey? PublicKey { get; set; }
-
-    public GuardianPublicKey() : base(nameof(GuardianPublicKey))
-    {
-    }
-}
-
-public class GuardianBackups : DatabaseRecord
-{
-    public string? KeyCeremonyId { get; set; }
-
-    public string? GuardianId { get; set; }
-
-    public string? DesignatedId { get; set; }
-
-    public ElectionPartialKeyBackup? Backup { get; set; }
-
-    public GuardianBackups() : base(nameof(GuardianBackups))
-    {
-    }
-}
-
-
-
 public partial class KeyCeremony : DatabaseRecord
 {
     public KeyCeremony(string name, int numberOfGuardians, int quorum, string admin) : base(nameof(KeyCeremony))
@@ -99,14 +46,3 @@ public partial class KeyCeremony : DatabaseRecord
     [ObservableProperty]
     private KeyCeremonyState _state;
 }
-
-
-//"guardians_joined": [],
-//"keys": [],
-//"guardians_keys": [],
-//"other_keys": [],
-//"backups": [],
-//"shared_backups": [],
-//"verifications": [],
-
-
